Replacing wagmin contract hooks with new hooks
