+++ services/client/src/hooks/index.ts
@@ -1,14 +1,11 @@
 export * from './useAuth';
 export * from './useAxiosInterceptors';
-export * from './useContractActivatePool';
-export * from './useContractCreatePool';
-export * from './useContractERC20Approve';
-export * from './useContractEscrowDeposit';
-export * from './useContractEscrowGetDepositsOf';
-export * from './useContractGetPool';
-export * from './useContractPoolURI';
 export * from './useLoadingModals';
 export * from './useSwitchNetwork';
 export * from './useWallet';
 export * from './useTokenDetails';
-export * from './useContractEscrowClaim';
+export * from './useContractEscrow';
+export * from './useContractPools';
+export * from './useEscrow';
+export * from './useHypePools';
+export * from './useMainnet';
