+++ services/client/src/pages/AddHypePool/AddHypePool.effects.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useRef, useState } from 'react';
-import { useContractCreatePool, WritePoolArgs } from '../../hooks/useContractCreatePool';
 import { ModalsActionsEnum, useModalsDispatch } from '../../context';
 import { HypePoolDetailsForm } from './DetailsForm';
 import { HypePoolRewardForm } from './RewardForm';
@@ -7,22 +6,13 @@ import { ethers } from 'ethers';
 import { useIpfsUpload } from '../../api/ipfs/useUploadIpfs';
 import { HypeProjectDetails } from '../../models';
 import { HypeImageUploadRef } from './DetailsForm/HypeImageUpload';
-import { NotificationType } from '../../utils';
-import { useHypePools } from '../../hooks/useHypePools';
+import { WritePoolArgs, useHypePools } from '../../hooks';
 
 export const useAddHypePoolEffects = () => {
   const dispatchModals = useModalsDispatch();
   const imageUploadRef = useRef<HypeImageUploadRef>(null);
 
-  const defaultContractArgs: WritePoolArgs = {
-    uri: null,
-    details: null,
-    rewards: null,
-  };
-  const resetWriteContract = (): void => {
-    setWritePoolArgs(defaultContractArgs);
-    setContractEnabled(false);
-  };
+
   const successCallback = (): void => {
     setCurrentStep(3);
   };
@@ -31,8 +21,6 @@ export const useAddHypePoolEffects = () => {
   };
 
   const { data: uploadedIpfsUrl, submitHandler } = useIpfsUpload();
-  const [writePoolArgs, setWritePoolArgs] = useState<WritePoolArgs>(defaultContractArgs);
-  const [contractEnabled, setContractEnabled] = useState<boolean>(false);
   const [createdPoolIndex, setCreatedPoolIndex] = useState<string>();
   const [currentStep, setCurrentStep] = useState<number>(1);
   const [ipfsUrl, setIpfsUrl] = useState<string>();
@@ -59,15 +47,6 @@ export const useAddHypePoolEffects = () => {
     endDate: 0,
   });
   const { createPool: testCreatePool } = useHypePools();
-
-  useContractCreatePool(
-    writePoolArgs,
-    contractEnabled,
-    resetWriteContract,
-    successCallback,
-    setCreatedPoolIndex,
-    setPoolTransaction,
-  );
 
   useEffect(() => {
     if (uploadedIpfsUrl) {
@@ -124,20 +103,7 @@ export const useAddHypePoolEffects = () => {
       rewards.impressionReward.toString().replace(',', '.'),
       rewards.tokenDecimals,
     );
-    setWritePoolArgs({
-      uri: ipfsUrl,
-      details,
-      rewards: {
-        ...rewards,
-        cap,
-        impressionReward,
-        tokenAddress:
-          rewards.tokenSymbol && rewards.tokenAddress ? rewards.tokenAddress : rewards.token,
-        endDate: 0,
-        startDate: 0,
-        duration: rewards.duration * 24 * 60 * 60,
-      },
-    });
+    
     const args: WritePoolArgs = {
       uri: ipfsUrl,
       details,
@@ -158,7 +124,6 @@ export const useAddHypePoolEffects = () => {
       setCreatedPoolIndex,
       setPoolTransaction,
     );
-    // setContractEnabled(true);
     setPoolReward(rewards);
   };
